Create list of number elements
